Translated integer based scene units to float based. Fixed dancing notes&staffs bug. Fixed crash when deleting the playing note.

+++ src/drawable/drawableclef.h
@@ -1,7 +1,7 @@
 /*!
 	Copyright (c) 2006-2007, Matevž Jekovec, Canorus development team
 	All Rights Reserved. See AUTHORS for a complete list of authors.
-	
+
 	Licensed under the GNU GENERAL PUBLIC LICENSE. See COPYING for details.
 */
 
@@ -15,12 +15,12 @@ class CADrawableStaff;
 
 class CADrawableClef : public CADrawableMusElement {
 public:
-	CADrawableClef(CAClef *clef, CADrawableStaff *drawableStaff, int x, int y);
-	
+	CADrawableClef(CAClef *clef, CADrawableStaff *drawableStaff, double x, double y);
+
 	void draw(QPainter *p, CADrawSettings s);
 	CADrawableClef *clone(CADrawableContext* newContext = 0);
 	inline CAClef *clef() { return (CAClef*)_musElement; }
-	
+
 	static const int CLEF_EIGHT_SIZE;
 };
