+++ src/drawable/drawablemark.h
@@ -1,7 +1,7 @@
 /*!
 	Copyright (c) 2007, Matevž Jekovec, Canorus development team
 	All Rights Reserved. See AUTHORS for a complete list of authors.
-	
+
 	Licensed under the GNU GENERAL PUBLIC LICENSE. See COPYING for details.
 */
 
@@ -18,21 +18,21 @@ class CADrawableNote;
 
 class CADrawableMark : public CADrawableMusElement {
 public:
-	CADrawableMark( CAMark *mark, CADrawableContext *drawableContext, int x, int y);
+	CADrawableMark( CAMark *mark, CADrawableContext *drawableContext, double x, double y);
 	virtual ~CADrawableMark();
-	
+
 	void draw( QPainter *p, CADrawSettings s );
 	CADrawableMark *clone( CADrawableContext* newContext = 0 );
 	inline CAMark *mark() { return static_cast<CAMark*>(musElement()); }
-	
+
 	inline void setRehersalMarkNumber( int n ) { _rehersalMarkNumber = n; }
 	inline int rehersalMarkNumber() { return _rehersalMarkNumber; }
-	
+
 	static QString fingerListToString( const QList<CAFingering::CAFingerNumber> list );
-	
+
 private:
-	static const int DEFAULT_TEXT_SIZE;
-	static const int DEFAULT_PIXMAP_SIZE;
+	static const double DEFAULT_TEXT_SIZE;
+	static const double DEFAULT_PIXMAP_SIZE;
 	CANote         *_tempoNote;
 	CADrawableNote *_tempoDNote;
 	QPixmap        *_pixmap;
